+++ src/main/java/wooteco/subway/repository/dao/LineDao.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dao;
+package wooteco.subway.repository.dao;
 
 import java.util.List;
