+++ src/main/java/wooteco/subway/service/LineService.java
@@ -5,8 +5,8 @@ import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 
-import wooteco.subway.dao.LineDao;
 import wooteco.subway.domain.line.Line;
+import wooteco.subway.repository.dao.LineDao;
 import wooteco.subway.service.dto.ServiceDtoAssembler;
 import wooteco.subway.service.dto.line.LineRequestDto;
 import wooteco.subway.service.dto.line.LineResponseDto;
