+++ src/test/java/wooteco/subway/repository/dao/FakeStationDao.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dao;
+package wooteco.subway.repository.dao;
 
 import java.util.ArrayList;
 import java.util.List;
